Fix: Remove Accounts import statements and separte Profile model from collection
