+++ .meteor/packages
@@ -6,17 +6,18 @@
 
 meteor-base@1.5.1             # Packages every Meteor app needs to have
 mobile-experience@1.1.0       # Packages for a great mobile UX
-mongo@1.15.0                   # The database Meteor supports right now
+mongo@1.16.0                   # The database Meteor supports right now
 
-standard-minifier-css@1.8.1   # CSS minifier run for production mode
-standard-minifier-js@2.8.0    # JS minifier run for production mode
+standard-minifier-css@1.8.2   # CSS minifier run for production mode
+standard-minifier-js@2.8.1    # JS minifier run for production mode
 es5-shim@4.8.0                # ECMAScript 5 compatibility for older browsers
 ecmascript@0.16.2              # Enable ECMAScript2015+ syntax in app code
 typescript@4.5.4              # Enable TypeScript syntax in .ts and .tsx modules
 shell-server@0.5.0            # Server-side component of the `meteor shell` command
 
 static-html@1.3.2             # Define static page content in .html files
-accounts-base@2.2.3
+vite:bundler
+accounts-base@2.2.4
 accounts-password@2.3.1
 check@1.3.1
 mdg:validated-method
@@ -32,4 +33,3 @@ jagi:astronomy
 xolvio:cleaner
 dburles:factory
 meteortesting:mocha
-vite:bundler
