+++ .meteor/release
@@ -1,1 +1,1 @@
-METEOR@2.7.3
+METEOR@2.8.0
