+++ .meteor/versions
@@ -16,26 +16,26 @@ boilerplate-generator@1.7.1
 caching-compiler@1.2.2
 caching-html-compiler@1.2.1
 callback-hook@1.4.0
-check@1.3.1
+check@1.3.2
 coffeescript@2.4.1
 coffeescript-compiler@2.4.1
 dburles:factory@1.1.0
-ddp@1.4.0
-ddp-client@2.5.0
+ddp@1.4.1
+ddp-client@2.6.1
 ddp-common@1.4.0
-ddp-rate-limiter@1.1.0
-ddp-server@2.5.0
-diff-sequence@1.1.1
+ddp-rate-limiter@1.1.1
+ddp-server@2.6.0
+diff-sequence@1.1.2
 dynamic-import@0.7.2
-ecmascript@0.16.2
+ecmascript@0.16.3
 ecmascript-runtime@0.8.0
 ecmascript-runtime-client@0.12.1
 ecmascript-runtime-server@0.11.0
-ejson@1.1.2
-email@2.2.1
+ejson@1.1.3
+email@2.2.2
 es5-shim@4.8.0
-fetch@0.1.1
-geojson-utils@1.0.10
+fetch@0.1.2
+geojson-utils@1.0.11
 hot-code-push@1.0.4
 html-tools@1.1.3
 htmljs@1.1.1
@@ -51,7 +51,7 @@ maka:rest@2.0.12
 matb33:collection-hooks@1.1.4
 mdg:validated-method@1.2.0
 mdg:validation-error@0.5.1
-meteor@1.10.0
+meteor@1.10.2
 meteor-base@1.5.1
 meteorhacks:async@1.0.0
 meteorhacks:ssr@2.2.0
@@ -60,26 +60,26 @@ meteortesting:mocha@2.0.3
 meteortesting:mocha-core@8.0.1
 minifier-css@1.6.1
 minifier-js@2.7.5
-minimongo@1.8.0
+minimongo@1.9.0
 mobile-experience@1.1.0
 mobile-status-bar@1.1.0
-modern-browsers@0.1.8
-modules@0.18.0
-modules-runtime@0.13.0
-mongo@1.15.0
+modern-browsers@0.1.9
+modules@0.19.0
+modules-runtime@0.13.1
+mongo@1.16.0
 mongo-decimal@0.1.3
 mongo-dev-server@1.1.0
 mongo-id@1.0.8
-npm-mongo@4.3.1
+npm-mongo@4.9.0
 observe-sequence@1.0.20
 ordered-dict@1.1.0
 peerlibrary:middleware@0.3.0
-promise@0.12.0
+promise@0.12.1
 raix:eventemitter@1.0.0
-random@1.2.0
+random@1.2.1
 rate-limit@1.0.9
 react-fast-refresh@0.2.3
-reactive-var@1.0.11
+reactive-var@1.0.12
 reload@1.3.1
 retry@1.1.0
 routepolicy@1.1.1
@@ -101,11 +101,11 @@ standard-minifier-js@2.8.1
 static-html@1.3.2
 templating-tools@1.2.2
 tmeasday:check-npm-versions@1.0.2
-tracker@1.2.0
+tracker@1.2.1
 typescript@4.5.4
-underscore@1.0.10
+underscore@1.0.11
 url@1.3.2
-vite:bundler@0.1.7
-webapp@1.13.1
-webapp-hashing@1.1.0
+vite:bundler@0.1.9
+webapp@1.13.2
+webapp-hashing@1.1.1
 xolvio:cleaner@0.4.0
