+++ imports/ui/mixins/accounts/profiles.ts
@@ -1,5 +1,6 @@
-import { ProfileCollection } from '../../../api/Profiles/Profile';
+
 import { defineComponent } from 'vue';
+import { ProfileCollection } from '/imports/api/Profiles/ProfileCollection';
 
 export default defineComponent({
 	meteor: {
