+++ package.json
@@ -21,7 +21,6 @@
     "fs-extra": "^10.0.0",
     "helmet": "^4.6.0",
     "iso-datestring-validator": "^2.2.0",
-    "jquery": "^3.6.0",
     "meteor-node-stubs": "^1.2.5",
     "mimetypes": "^0.1.1",
     "mitt": "^3.0.0",
@@ -51,7 +50,6 @@
     "@types/sinon": "^10.0.4",
     "@types/webfontloader": "^1.6.35",
     "@vitejs/plugin-vue": "^3.2.0",
-    "autoprefixer": "^10.4.13",
     "chai": "^4.3.4",
     "faker": "^5.5.3",
     "husky": "^7.0.2",
